backend/chatgpt/pyproject.toml
[tool.poetry]
name = "projectweb"
version = "0.1.0"
description = ""
authors = ["syrlybajramazan <syrlybai056@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
rest-framework-simplejwt = "^0.0.2"
pyjwt = "^2.8.0"
openai = "0.28"
django-environ = "^0.11.2"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
